requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-dev.in
#
--no-binary psycopg2

appdirs==1.4.4            # via black, virtualenv
attrs==20.2.0             # via flake8-bugbear, pytest
autoflake==1.4            # via -r requirements-dev.in
autopep8==1.5.4           # via -r requirements-dev.in
backcall==0.2.0           # via ipython
black==20.8b1             # via -r requirements-dev.in
certifi==2020.6.20        # via -c requirements.txt, requests
cfgv==3.2.0               # via pre-commit
chardet==3.0.4            # via -c requirements.txt, requests
click==7.1.2              # via black, pip-tools
coverage==5.3             # via pytest-cov
decorator==4.4.2          # via ipython
distlib==0.3.1            # via virtualenv
ecdsa==0.14.1             # via -c requirements.txt, python-jose
factory-boy==3.1.0        # via -r requirements-dev.in, pytest-factoryboy
faker==4.14.0             # via factory-boy
fastdiff==0.2.0           # via snapshottest
filelock==3.0.12          # via virtualenv
flake8-bugbear==20.1.4    # via -r requirements-dev.in
flake8-polyfill==1.0.2    # via pep8-naming
flake8==3.8.4             # via -r requirements-dev.in, flake8-bugbear, flake8-polyfill
freezegun==1.0.0          # via -r requirements-dev.in
identify==1.5.6           # via pre-commit
idna==2.10                # via -c requirements.txt, requests
importmagic==0.1.7        # via -r requirements-dev.in
inflection==0.5.1         # via pytest-factoryboy
iniconfig==1.1.1          # via pytest
ipython-genutils==0.2.0   # via traitlets
ipython==7.19.0           # via -r requirements-dev.in
isort==5.6.4              # via -r requirements-dev.in
jedi==0.17.2              # via -r requirements-dev.in, ipython
mccabe==0.6.1             # via flake8
mypy-extensions==0.4.3    # via black
nodeenv==1.5.0            # via pre-commit
packaging==20.4           # via pytest
parso==0.7.1              # via jedi
pathspec==0.8.0           # via black
pep8-naming==0.11.1       # via -r requirements-dev.in
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.3.1          # via -r requirements-dev.in
pluggy==0.13.1            # via pytest
pre-commit==2.8.2         # via -r requirements-dev.in
prompt-toolkit==3.0.8     # via ipython
ptyprocess==0.6.0         # via pexpect
py==1.9.0                 # via pytest
pyasn1==0.4.8             # via -c requirements.txt, python-jose, rsa
pycodestyle==2.6.0        # via autopep8, flake8
pydocstyle==5.1.1         # via -r requirements-dev.in
pyflakes==2.2.0           # via autoflake, flake8
pygments==2.7.2           # via ipython
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.1        # via -r requirements-dev.in
pytest-django==4.1.0      # via -r requirements-dev.in
pytest-factoryboy==2.0.3  # via -r requirements-dev.in
pytest-lazy-fixture==0.6.3  # via -r requirements-dev.in
pytest-mock==3.5.0        # via -r requirements-dev.in
pytest==6.1.2             # via -r requirements-dev.in, pytest-cov, pytest-django, pytest-factoryboy, pytest-lazy-fixture, pytest-mock
python-dateutil==2.8.1    # via faker, freezegun
python-jose==3.2.0        # via -c requirements.txt, -r requirements-dev.in
pyyaml==5.3.1             # via -c requirements.txt, pre-commit
regex==2020.10.28         # via black
requests-mock==1.8.0      # via -r requirements-dev.in
requests==2.24.0          # via -c requirements.txt, requests-mock
rope==0.18.0              # via -r requirements-dev.in
rsa==4.6                  # via -c requirements.txt, python-jose
six==1.15.0               # via -c requirements.txt, ecdsa, packaging, pip-tools, python-dateutil, python-jose, requests-mock, snapshottest, virtualenv
snapshottest==0.6.0       # via -r requirements-dev.in
snowballstemmer==2.0.0    # via pydocstyle
termcolor==1.1.0          # via snapshottest
text-unidecode==1.3       # via faker
toml==0.10.2              # via autopep8, black, pre-commit, pytest
traitlets==5.0.5          # via ipython
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
urllib3==1.25.11          # via -c requirements.txt, requests
virtualenv==20.1.0        # via pre-commit
wasmer==0.4.1             # via fastdiff
wcwidth==0.2.5            # via prompt-toolkit
yapf==0.30.0              # via -r requirements-dev.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
